+++ HISTORY.md
@@ -10,11 +10,13 @@
 * **Notable Fixes:**
     * strings library: fixed lowercase(number), thanks, Brian and Simon!
     * swapped "lists" and "other" in the make-a-block category menu, thanks Brian, for the report!
+    * don't let deleted sprites clone
 * **Documentation Updates:**
 * **Translation Updates:**
 
 ### 2022-02-21
 * threads: optimized hyper-if/else to skip repeated evaluation of literal true/false cases
+* objects: prevent deleted sprites from cloning
 
 ### 2022-02-18
 * new MQTT extension and library, thanks, Simon and Xavier!